core: opts: move opts from core module to bpfilter
opts.{c,h} are daemon configuration options, they are not meant to be
used by bfcli or libbpfilter.

This change move opts.{c,h} to src/bpfilter, where they belong.
