+++ src/bpfilter/cgen/program.c
@@ -38,6 +38,7 @@
 #include "bpfilter/cgen/tc.h"
 #include "bpfilter/cgen/xdp.h"
 #include "bpfilter/ctx.h"
+#include "bpfilter/opts.h"
 #include "core/bpf.h"
 #include "core/btf.h"
 #include "core/chain.h"
@@ -51,7 +52,6 @@
 #include "core/logger.h"
 #include "core/marsh.h"
 #include "core/matcher.h"
-#include "core/opts.h"
 #include "core/rule.h"
 #include "core/set.h"
 #include "core/verdict.h"
