+++ src/bpfilter/xlate/ipt/ipt.c
@@ -19,6 +19,7 @@
 #include "bpfilter/cgen/cgen.h"
 #include "bpfilter/cgen/program.h"
 #include "bpfilter/ctx.h"
+#include "bpfilter/opts.h"
 #include "bpfilter/xlate/front.h"
 #include "bpfilter/xlate/ipt/dump.h"
 #include "bpfilter/xlate/ipt/helpers.h"
@@ -32,7 +33,6 @@
 #include "core/logger.h"
 #include "core/marsh.h"
 #include "core/matcher.h"
-#include "core/opts.h"
 #include "core/request.h"
 #include "core/response.h"
 #include "core/rule.h"
